requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pypdf
ollama
python-multipart
python-dotenv
groq